docs(readme): Add section summarizing 7 rules for writing good commit messages
In this commit, both the title and the body were added.
The title and body are separated by one blank line.
Previously, the title started with a lowercase letter, but from now on,
I will make it a habit to start the title with an uppercase letter.
I also followed the other conventions such as breaking lines every 72
characters in the body